requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-CORS>=4.0.0
Flask-JWT-Extended>=4.5.0
Flask-Caching>=2.0.0
Flask-Compress>=1.14
Werkzeug>=2.3.0
SQLAlchemy>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0
psycopg2-binary>=2.9.0
redis>=5.0.0